refactor dataset functions into dataset module
