+++ openml/__init__.py
@@ -16,11 +16,12 @@ In particular, this module implemts a python interface for the
 """
 
 from .apiconnector import APIConnector
-from .entities.dataset import OpenMLDataset
+from .datasets import OpenMLDataset
+from . import datasets
 from .entities.run import OpenMLRun
 from .entities.split import OpenMLSplit
 
 
 __version__ = "0.2.1"
 
-__all__ = ['APIConnector', 'OpenMLDataset', 'OpenMLRun', 'OpenMLSplit']
+__all__ = ['APIConnector', 'OpenMLDataset', 'OpenMLRun', 'OpenMLSplit', 'datasets']
