+++ openml/datasets/__init__.py
@@ -1,0 +1,2 @@
+from .functions import *
+from .dataset import OpenMLDataset
