+++ tests/test_apiconnector.py
@@ -13,7 +13,7 @@ if sys.version_info[0] >= 3:
     from urllib.error import URLError
 else:
     import mock
-    from urllib import urlencode, urlopen
+    from urllib import urlencode
     from urllib2 import URLError, urlopen
 
 
@@ -51,13 +51,9 @@ class TestAPIConnector(unittest.TestCase):
         except:
             apikey = None
 
-        try:
-            travis = os.environ['TRAVIS']
-            if apikey is None:
-                raise Exception('Running on travis-ci, but no environment '
-                                'variable OPENMLAPIKEY found.')
-        except:
-            pass
+        if "TRAVIS" in os.environ and apikey is None:
+            raise Exception('Running on travis-ci, but no environment '
+                            'variable OPENMLAPIKEY found.')
 
         self.connector = APIConnector(cache_directory=self.workdir,
                                       apikey=apikey)
@@ -72,40 +68,6 @@ class TestAPIConnector(unittest.TestCase):
     def test_parse_config(self):
         raise Exception()
 
-    ############################################################################
-    # Test all local stuff
-    def test_get_cached_datasets(self):
-        workdir = os.path.dirname(os.path.abspath(__file__))
-        workdir = os.path.join(workdir, "files")
-        connector = APIConnector(cache_directory=workdir)
-        datasets = connector.get_cached_datasets()
-        self.assertIsInstance(datasets, dict)
-        self.assertEqual(len(datasets), 2)
-        self.assertIsInstance(list(datasets.values())[0], OpenMLDataset)
-
-    def test_get_cached_dataset(self):
-        workdir = os.path.dirname(os.path.abspath(__file__))
-        workdir = os.path.join(workdir, "files")
-
-        with mock.patch.object(APIConnector, "_perform_api_call") as api_mock:
-            api_mock.return_value = 400, \
-                """<oml:authenticate xmlns:oml = "http://openml.org/openml">
-                <oml:session_hash>G9MPPN114ZCZNWW2VN3JE9VF1FMV8Y5FXHUDUL4P</oml:session_hash>
-                <oml:valid_until>2014-08-13 20:01:29</oml:valid_until>
-                <oml:timezone>Europe/Berlin</oml:timezone>
-                </oml:authenticate>"""
-
-            connector = APIConnector(cache_directory=workdir)
-            dataset = connector.get_cached_dataset(2)
-            self.assertIsInstance(dataset, OpenMLDataset)
-            self.assertTrue(connector._perform_api_call.is_called_once())
-
-    def test_get_chached_dataset_description(self):
-        workdir = os.path.dirname(os.path.abspath(__file__))
-        workdir = os.path.join(workdir, "files")
-        connector = APIConnector(cache_directory=workdir)
-        description = connector._get_cached_dataset_description(2)
-        self.assertIsInstance(description, dict)
 
     @unittest.skip("Not implemented yet.")
     def test_get_cached_tasks(self):
@@ -126,71 +88,6 @@ class TestAPIConnector(unittest.TestCase):
     ############################################################################
     # Test all remote stuff
 
-    ############################################################################
-    # Datasets
-    def test_get_dataset_list(self):
-        # We can only perform a smoke test here because we test on dynamic
-        # data from the internet...
-        datasets = self.connector.get_dataset_list()
-        # 1087 as the number of datasets on openml.org
-        self.assertTrue(len(datasets) >= 1087)
-        for dataset in datasets:
-            self.assertEqual(type(dataset), dict)
-            self.assertGreaterEqual(len(dataset), 2)
-            self.assertIn('did', dataset)
-            self.assertIsInstance(dataset['did'], int)
-            self.assertIn('status', dataset)
-            self.assertTrue(is_string(dataset['status']))
-            self.assertIn(dataset['status'], ['in_preparation', 'active',
-                                              'deactivated'])
-
-    @unittest.skip("Not implemented yet.")
-    def test_datasets_active(self):
-        raise NotImplementedError()
-
-    def test_download_datasets(self):
-        dids = [1, 2]
-        datasets = self.connector.download_datasets(dids)
-        self.assertEqual(len(datasets), 2)
-        self.assertTrue(os.path.exists(os.path.join(
-            self.connector.dataset_cache_dir, "1", "description.xml")))
-        self.assertTrue(os.path.exists(os.path.join(
-            self.connector.dataset_cache_dir, "2", "description.xml")))
-        self.assertTrue(os.path.exists(os.path.join(
-            self.connector.dataset_cache_dir, "1", "dataset.arff")))
-        self.assertTrue(os.path.exists(os.path.join(
-            self.connector.dataset_cache_dir, "2", "dataset.arff")))
-
-    def test_download_dataset(self):
-        dataset = self.connector.download_dataset(1)
-        self.assertEqual(type(dataset), OpenMLDataset)
-        self.assertEqual(dataset.name, 'anneal')
-        self.assertTrue(os.path.exists(os.path.join(
-            self.connector.dataset_cache_dir, "1", "description.xml")))
-        self.assertTrue(os.path.exists(os.path.join(
-            self.connector.dataset_cache_dir, "1", "dataset.arff")))
-
-    def test_download_rowid(self):
-        # Smoke test which checks that the dataset has the row-id set correctly
-        did = 164
-        dataset = self.connector.download_dataset(did)
-        self.assertEqual(dataset.row_id_attribute, 'instance')
-
-    def test_download_dataset_description(self):
-        # Only a smoke test, I don't know exactly how to test the URL
-        # retrieval and "caching"
-        description = self.connector.download_dataset_description(2)
-        self.assertIsInstance(description, dict)
-
-    def test_download_dataset_features(self):
-        # Only a smoke check
-        features = self.connector.download_dataset_features(2)
-        self.assertIsInstance(features, dict)
-
-    def test_download_dataset_qualities(self):
-        # Only a smoke check
-        qualities = self.connector.download_dataset_qualities(2)
-        self.assertIsInstance(qualities, dict)
 
     ############################################################################
     # Tasks
@@ -277,36 +174,6 @@ class TestAPIConnector(unittest.TestCase):
         for flow in flows:
             check_flow(flow)
 
-    def test_upload_dataset(self):
-
-        dataset = self.connector.download_dataset(3)
-        file_path = os.path.join(self.connector.dataset_cache_dir, "3", "dataset.arff")
-
-        description = """ <oml:data_set_description xmlns:oml="http://openml.org/openml">
-                        <oml:name>anneal</oml:name>
-                        <oml:version>1</oml:version>
-                        <oml:description>test</oml:description>
-                        <oml:format>ARFF</oml:format>
-                        <oml:licence>Public</oml:licence>
-                        <oml:default_target_attribute>class</oml:default_target_attribute>
-                        <oml:md5_checksum></oml:md5_checksum>
-                        </oml:data_set_description>
-                         """
-        return_code, dataset_xml = self.connector.upload_dataset(description, file_path)
-        self.assertEqual(return_code, 200)
-
-    def test_upload_dataset_with_url(self):
-
-        description = """ <oml:data_set_description xmlns:oml="http://openml.org/openml">
-                        <oml:name>UploadTestWithURL</oml:name>
-                        <oml:version>1</oml:version>
-                        <oml:description>test</oml:description>
-                        <oml:format>ARFF</oml:format>
-                        <oml:url>http://expdb.cs.kuleuven.be/expdb/data/uci/nominal/iris.arff</oml:url>
-                        </oml:data_set_description>
-                         """
-        return_code, dataset_xml = self.connector.upload_dataset(description)
-        self.assertEqual(return_code, 200)
 
     def test_upload_flow(self):
         description = '''<oml:flow xmlns:oml="http://openml.org/openml"><oml:name>Test</oml:name><oml:description>description</oml:description> </oml:flow>'''
